requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
dnspython==2.8.0
email-validator==2.3.0
frozenlist==1.8.0
idna==3.11
lxml==6.0.2
magic-filter==1.0.12
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-whois==0.9.6
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
validators==0.35.0
yarl==1.23.0
